Extracted EventData, following pattern of ProfileData to support one source of truth for how event participation is queried.
